Less messy.
